Move the functions and classes from temporary `byTbp.js` to their rightful place. Add a constants.js module.
https://github.com/metasfresh/metasfresh-e2e/issues/133
