+++ cypress/support/commands/form.js
@@ -1,4 +1,4 @@
-import { confirmCalendarDay } from '../functions';
+import {confirmCalendarDay} from '../functions';
 
 // thx to https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/String/replace
 function removeSubstringsWithCurlyBrackets(stringValue) {
@@ -11,7 +11,7 @@ function removeSubstringsWithCurlyBrackets(stringValue) {
 }
 
 Cypress.Commands.add('clearField', (fieldName, modal) => {
-  describe('Clear field', function() {
+  describe('Clear field', function () {
     cy.log(`clearField - fieldName=${fieldName}; modal=${modal}`);
 
     const path = createFieldPath(fieldName, modal);
@@ -22,7 +22,7 @@ Cypress.Commands.add('clearField', (fieldName, modal) => {
 });
 
 Cypress.Commands.add('getStringFieldValue', (fieldName, modal) => {
-  describe('Get field value', function() {
+  describe('Get field value', function () {
     cy.log(`getStringFieldValue - fieldName=${fieldName}; modal=${modal}`);
 
     const path = createFieldPath(fieldName, modal);
@@ -34,7 +34,7 @@ Cypress.Commands.add('getStringFieldValue', (fieldName, modal) => {
 });
 
 Cypress.Commands.add('assertFieldNotShown', (fieldName, modal) => {
-  describe('Assert that a vield is not shown', function() {
+  describe('Assert that a vield is not shown', function () {
     cy.log(`assertFieldNotShown - fieldName=${fieldName}; modal=${modal}`);
 
     const path = createFieldPath(fieldName, modal);
@@ -51,7 +51,7 @@ function createFieldPath(fieldName, modal) {
 }
 
 Cypress.Commands.add('getCheckboxValue', (fieldName, modal) => {
-  describe('Get field value', function() {
+  describe('Get field value', function () {
     cy.log(`getCheckboxValue - fieldName=${fieldName}; modal=${modal}`);
 
     const path = createFieldPath(fieldName, modal);
@@ -66,7 +66,7 @@ Cypress.Commands.add('getCheckboxValue', (fieldName, modal) => {
 });
 
 Cypress.Commands.add('resetListValue', (fieldName, modal, rewriteUrl = null) => {
-  describe('Get field value', function() {
+  describe('Get field value', function () {
     cy.log(`resetListValue - fieldName=${fieldName}; modal=${modal}`);
 
     const patchUrlPattern = rewriteUrl || '/rest/api/window/.*[^/][^N][^E][^W]$';
@@ -85,7 +85,7 @@ Cypress.Commands.add('resetListValue', (fieldName, modal, rewriteUrl = null) => 
 });
 
 Cypress.Commands.add('clickOnIsActive', modal => {
-  describe('Click on the IsActive slider', function() {
+  describe('Click on the IsActive slider', function () {
     const path = createFieldPath('IsActive', modal);
 
     cy.get(path)
@@ -98,7 +98,7 @@ Cypress.Commands.add('clickOnIsActive', modal => {
  * @param modal - use true if the field is in a modal overlay; required if the underlying window has a field with the same name
  */
 Cypress.Commands.add('clickOnCheckBox', (fieldName, expectedPatchValue, modal, rewriteUrl = null) => {
-  describe('Click on a checkbox field', function() {
+  describe('Click on a checkbox field', function () {
     cy.log(`clickOnCheckBox - fieldName=${fieldName}`);
 
     const patchUrlPattern = rewriteUrl || '/rest/api/window/.*[^/][^N][^E][^W]$';
@@ -113,7 +113,7 @@ Cypress.Commands.add('clickOnCheckBox', (fieldName, expectedPatchValue, modal, r
 
     cy.get(path)
       .find('.input-checkbox-tick')
-      .click({ force: true }) // we don't care if the checkbox scrolled out of view
+      .click({force: true}) // we don't care if the checkbox scrolled out of view
       .waitForFieldValue(`@${patchCheckBoxAliasName}`, fieldName, expectedPatchValue);
   });
 });
@@ -144,7 +144,7 @@ Cypress.Commands.add('selectDateViaPicker', (fieldName, modal) => {
  * @param {boolean} noRequest - if set to true, don't wait for the response from the server
  */
 Cypress.Commands.add('writeIntoStringField', (fieldName, stringValue, modal, rewriteUrl, noRequest) => {
-  describe('Enter value into string field', function() {
+  describe('Enter value into string field', function () {
     const aliasName = `writeIntoStringField-${new Date().getTime()}`;
     const expectedPatchValue = removeSubstringsWithCurlyBrackets(stringValue);
     // in the default pattern we want to match URLs that do *not* end with "/NEW"
@@ -180,7 +180,7 @@ Cypress.Commands.add('writeIntoStringField', (fieldName, stringValue, modal, rew
  * @param {string} rewriteUrl - use custom url for the request
  */
 Cypress.Commands.add('writeIntoTextField', (fieldName, stringValue, modal, rewriteUrl) => {
-  describe('Enter value into text field', function() {
+  describe('Enter value into text field', function () {
     cy.log(`writeIntoTextField - fieldName=${fieldName}; stringValue=${stringValue}; modal=${modal}`);
 
     const aliasName = `writeIntoTextField-${new Date().getTime()}`;
@@ -209,7 +209,7 @@ Cypress.Commands.add('writeIntoTextField', (fieldName, stringValue, modal, rewri
 Cypress.Commands.add(
   'writeIntoLookupListField',
   (fieldName, partialValue, expectedListValue, typeList = false, modal = false, rewriteUrl = null) => {
-    describe('Enter value into lookup list field', function() {
+    describe('Enter value into lookup list field', function () {
       let path = `#lookup_${fieldName}`;
       if (modal) {
         path = `.panel-modal ${path}`;
@@ -254,7 +254,7 @@ Cypress.Commands.add(
  * @param {boolean} skipRequest - if set to true, cypress won't expect a request to the server and won't wait for it
  */
 Cypress.Commands.add('selectInListField', (fieldName, listValue, modal, rewriteUrl = null, skipRequest) => {
-  describe('Select value in list field', function() {
+  describe('Select value in list field', function () {
     cy.log(`selectInListField - fieldName=${fieldName}; listValue=${listValue}; modal=${modal}`);
 
     const patchListFieldAliasName = `patchListField-${new Date().getTime()}`;
@@ -289,7 +289,7 @@ Cypress.Commands.add('selectInListField', (fieldName, listValue, modal, rewriteU
  * @param {boolean} modal - use true, if the field is in a modal overlay; requered if the underlying window has a field with the same name
  */
 Cypress.Commands.add('selectNthInListField', (fieldName, index, modal) => {
-  describe('Select n-th option in list field', function() {
+  describe('Select n-th option in list field', function () {
     cy.log(`selectNthInListField - fieldName=${fieldName}; index=${index}; modal=${modal}`);
 
     const path = createFieldPath(fieldName, modal);
@@ -301,6 +301,32 @@ Cypress.Commands.add('selectNthInListField', (fieldName, index, modal) => {
       for (let i = 0; i < options.length; i += 1) {
         if (i === index) {
           cy.get(options[i]).click();
+        }
+      }
+    });
+  });
+});
+
+
+Cypress.Commands.add('setCheckBoxValue', (fieldName, isChecked, modal = false, rewriteUrl = null) => {
+  describe(`Set the Checkbox value ${fieldName} to ${isChecked}`, function () {
+
+    // the expected value is the same as the checked state
+    // (used only for verification if the checkbox has the correct value)
+    const expectedPatchValue = isChecked;
+
+    cy.getCheckboxValue(fieldName, modal).then(theCheckboxValue => {
+      if (isChecked) {
+        if (theCheckboxValue) {
+          // Nothing to do, already checked
+        } else {
+          cy.clickOnCheckBox(fieldName, expectedPatchValue, modal, rewriteUrl);
+        }
+      } else {
+        if (theCheckboxValue) {
+          cy.clickOnCheckBox(fieldName, expectedPatchValue, modal, rewriteUrl);
+        } else {
+          // Nothing to do, already unchecked
         }
       }
     });
