+++ cypress/support/commands/test.js
@@ -1,5 +1,8 @@
+import {getLanguageSpecific} from "../utils/utils";
+import {DocumentStatusKey} from "../utils/constants";
+
 Cypress.Commands.add('editAddress', (fieldName, addressFunction) => {
-  describe(`Select ${fieldName}'s address-button and invoke the given function`, function() {
+  describe(`Select ${fieldName}'s address-button and invoke the given function`, function () {
     cy.server();
     cy.route('POST', '/rest/api/address').as('postAddress');
 
@@ -29,7 +32,7 @@ Cypress.Commands.add('editAddress', (fieldName, addressFunction) => {
  * @param waitBeforePress if truthy, call cy.wait with the given parameter first
  */
 Cypress.Commands.add('pressStartButton', waitBeforePress => {
-  describe("Press an overlay's start-button", function() {
+  describe("Press an overlay's start-button", function () {
     if (waitBeforePress) {
       cy.wait(waitBeforePress);
     }
@@ -46,7 +49,7 @@ Cypress.Commands.add('pressStartButton', waitBeforePress => {
 });
 
 Cypress.Commands.add('processDocument', (action, expectedStatus) => {
-  describe('Execute a doc action', function() {
+  describe('Execute a doc action', function () {
     cy.log(`Execute doc action ${action}`);
 
     cy.get('.form-field-DocAction .meta-dropdown-toggle').click();
@@ -59,7 +62,7 @@ Cypress.Commands.add('processDocument', (action, expectedStatus) => {
       .click();
     // .click({ force: true }) // force is needed in some cases with chrome71 (IDK why, to the naked eye the action seems to be visible)
 
-    cy.get('.indicator-pending', { timeout: 10000 }).should('not.exist');
+    cy.get('.indicator-pending', {timeout: 10000}).should('not.exist');
     if (expectedStatus) {
       cy.log(`Verify that the doc status is now ${expectedStatus}`);
       cy.get('.meta-dropdown-toggle .tag-success').contains(expectedStatus);
@@ -68,7 +71,7 @@ Cypress.Commands.add('processDocument', (action, expectedStatus) => {
 });
 
 Cypress.Commands.add('openAdvancedEdit', () => {
-  describe('Open the advanced edit overlay via ALT+E shortcut', function() {
+  describe('Open the advanced edit overlay via ALT+E shortcut', function () {
     cy.get('body').type('{alt}E');
     cy.get('.panel-modal').should('exist');
   });
@@ -80,7 +83,7 @@ Cypress.Commands.add('openAdvancedEdit', () => {
  * @param waitBeforePress if truthy, call cy.wait with the given parameter first
  */
 Cypress.Commands.add('pressDoneButton', waitBeforePress => {
-  describe("Press an overlay's done-button", function() {
+  describe("Press an overlay's done-button", function () {
     if (waitBeforePress) {
       cy.wait(waitBeforePress);
     }
@@ -98,13 +101,13 @@ Cypress.Commands.add('pressDoneButton', waitBeforePress => {
       .should('exist')
       .click();
 
-    cy.get('.panel-modal', { timeout: 10000 }) // wait up to 10 secs for the modal to appear
+    cy.get('.panel-modal', {timeout: 10000}) // wait up to 10 secs for the modal to appear
       .should('not.exist');
   });
 });
 
 Cypress.Commands.add('pressAddNewButton', (includedDocumentIdAliasName = 'newIncludedDocumentId') => {
-  describe("Press table's add-new-record-button", function() {
+  describe("Press table's add-new-record-button", function () {
     cy.server();
     // window/<windowId>/<rootDocumentId>/<tabId>/NEW
     cy.route('PATCH', new RegExp('/rest/api/window/[^/]+/[^/]+/[^/]+/NEW$')).as('patchNewIncludedDocument');
@@ -116,7 +119,7 @@ Cypress.Commands.add('pressAddNewButton', (includedDocumentIdAliasName = 'newInc
       .click()
       .wait('@patchNewIncludedDocument')
       .then(xhr => {
-        return { documentId: xhr.response.body[0].rowId };
+        return {documentId: xhr.response.body[0].rowId};
       })
       .as(includedDocumentIdAliasName);
 
@@ -129,7 +132,7 @@ Cypress.Commands.add('pressAddNewButton', (includedDocumentIdAliasName = 'newInc
  * @param waitBeforePress if truthy, call cy.wait with the given parameter first
  */
 Cypress.Commands.add('pressBatchEntryButton', waitBeforePress => {
-  describe("Press table's batch-entry-record-button", function() {
+  describe("Press table's batch-entry-record-button", function () {
     if (waitBeforePress) {
       cy.wait(waitBeforePress);
     }
@@ -141,5 +144,16 @@ Cypress.Commands.add('pressBatchEntryButton', waitBeforePress => {
       .click();
 
     cy.get('.quick-input-container').should('exist');
+  });
+});
+
+
+Cypress.Commands.add('expectDocumentStatus', (expectedDocumentStatus) => {
+  describe(`Expect specific document status`, function () {
+    cy.fixture('misc/misc_dictionary.json').then(miscDictionaryJson => {
+      const expectedTrl = getLanguageSpecific(miscDictionaryJson, expectedDocumentStatus);
+      const documentTag = DocumentStatusKey[`_tag_${expectedDocumentStatus}`];
+      cy.get(`.meta-dropdown-toggle ${documentTag}`).contains(expectedTrl);
+    });
   });
 });
